+++ packages/netsuite-adapter/test/adapter_creator.test.ts
@@ -482,110 +482,6 @@ describe('NetsuiteAdapter creator', () => {
           })
         ).toThrow()
       })
-
-      describe('validateClientConfig', () => {
-        describe('validateMaxInstancesPerType', () => {
-          it('should validate maxInstancesPerType is the correct object with valid NS types', () => {
-            const invalidConfig = new InstanceElement(
-              ElemID.CONFIG_NAME,
-              adapter.configType as ObjectType,
-              {
-                client: {
-                  maxInstancesPerType: [{ name: 'customsegment', limit: 3 }],
-                },
-              }
-            )
-            expect(
-              () => adapter.operations({
-                credentials,
-                config: invalidConfig,
-                getElemIdFunc: mockGetElemIdFunc,
-                elementsSource: buildElementsSourceFromElements([]),
-              })
-            ).not.toThrow()
-          })
-
-          it('should validate also customrecordtype instances', () => {
-            const invalidConfig = new InstanceElement(
-              ElemID.CONFIG_NAME,
-              adapter.configType as ObjectType,
-              {
-                client: {
-                  maxInstancesPerType: [{ name: 'customrecord_For*', limit: 3 }],
-                },
-              }
-            )
-            expect(
-              () => adapter.operations({
-                credentials,
-                config: invalidConfig,
-                getElemIdFunc: mockGetElemIdFunc,
-                elementsSource: buildElementsSourceFromElements([]),
-              })
-            ).not.toThrow()
-          })
-
-          it('should throw if maxInstancesPerType has an invalid regex', () => {
-            const invalidConfig = new InstanceElement(
-              ElemID.CONFIG_NAME,
-              adapter.configType as ObjectType,
-              {
-                client: {
-                  maxInstancesPerType: [{ name: 'custom(.*', limit: 3 }],
-                },
-              }
-            )
-            expect(
-              () => adapter.operations({
-                credentials,
-                config: invalidConfig,
-                getElemIdFunc: mockGetElemIdFunc,
-                elementsSource: buildElementsSourceFromElements([]),
-              })
-            ).toThrow()
-          })
-
-          it('should throw if maxInstancesPerType is the wrong object', () => {
-            const invalidConfig = new InstanceElement(
-              ElemID.CONFIG_NAME,
-              adapter.configType as ObjectType,
-              {
-                client: {
-                  maxInstancesPerType: [{ wrong_name: 'customsegment', limit: 3 }],
-                },
-              }
-            )
-            expect(
-              () => adapter.operations({
-                credentials,
-                config: invalidConfig,
-                getElemIdFunc: mockGetElemIdFunc,
-                elementsSource: buildElementsSourceFromElements([]),
-              })
-            ).toThrow()
-          })
-
-          it('should throw if maxInstancesPerType is the correct object with invalid NS types', () => {
-            const invalidConfig = new InstanceElement(
-              ElemID.CONFIG_NAME,
-              adapter.configType as ObjectType,
-              {
-                client: {
-                  maxInstancesPerType: [{ name: 'not_supported_type', limit: 3 }],
-                },
-              }
-            )
-            expect(
-              () => adapter.operations({
-                credentials,
-                config: invalidConfig,
-                getElemIdFunc: mockGetElemIdFunc,
-                elementsSource: buildElementsSourceFromElements([]),
-              })
-            ).toThrow()
-          })
-        })
-      })
     })
 
     describe('deploy params', () => {
