+++ packages/netsuite-adapter/test/client/sdf_client.test.ts
@@ -493,9 +493,7 @@ describe('sdf client', () => {
         .toHaveBeenNthCalledWith(numberOfExecuteActions, deleteAuthIdCommandMatcher)
     })
 
-    // SALTO-3042 Enable test after full deployment
-    // eslint-disable-next-line jest/no-disabled-tests
-    it.skip('should exclude types with too many instances', async () => {
+    it('should NOT exclude types with too many instances', async () => {
       mockExecuteAction.mockImplementation(context => {
         const ids = [
           { type: 'addressForm', scriptId: 'a' },
@@ -523,19 +521,22 @@ describe('sdf client', () => {
       await client.getCustomObjects(typeNames, typeNamesQuery)
       // createProject & setupAccount & listObjects & 1*importObjects & deleteAuthId
       const numberOfExecuteActions = 6
-      expect(mockExecuteAction).toHaveBeenCalledTimes(numberOfExecuteActions)
+      expect(mockExecuteAction).toHaveBeenCalledTimes(numberOfExecuteActions + 1)
       expect(mockExecuteAction).toHaveBeenNthCalledWith(1, createProjectCommandMatcher)
       expect(mockExecuteAction).toHaveBeenNthCalledWith(2, saveTokenCommandMatcher)
       expect(mockExecuteAction).toHaveBeenNthCalledWith(3, listObjectsCommandMatcher)
       expect(mockExecuteAction).toHaveBeenNthCalledWith(4, importObjectsCommandMatcher)
       expect(mockExecuteAction.mock.calls[3][0].arguments).toEqual(expect.objectContaining({
+        type: 'addressForm',
+      }))
+      expect(mockExecuteAction.mock.calls[3 + 1][0].arguments).toEqual(expect.objectContaining({
         type: 'advancedpdftemplate',
         scriptid: 'd',
       }))
 
-      expect(mockExecuteAction).toHaveBeenNthCalledWith(5, importConfigurationCommandMatcher)
+      expect(mockExecuteAction).toHaveBeenNthCalledWith(5 + 1, importConfigurationCommandMatcher)
       expect(mockExecuteAction)
-        .toHaveBeenNthCalledWith(numberOfExecuteActions, deleteAuthIdCommandMatcher)
+        .toHaveBeenNthCalledWith(numberOfExecuteActions + 1, deleteAuthIdCommandMatcher)
     })
 
     it('should succeed', async () => {
